Move apps from experimental settings to debug
