+++ CHANGELOG.md
@@ -36,6 +36,7 @@
 - [FIX] Wait for disconnect while update
 - [FIX] Fix mfkey32
 - [FIX] Use CurrentActivityHolder for get activity in updater(replace Context cast)
+- [FIX] Move apps from experimental settings to debug
 - [REFACTOR] Migrate bottom bar to compose navigation
 - [REFACTOR] Bump Android Gradle Plugin
 - [REFACTOR] Fix detekt compose issues
