+++ components/settings/impl/src/main/java/com/flipperdevices/settings/impl/composable/category/DebugCategory.kt
@@ -92,4 +92,10 @@ private fun DebugCategoryItems(
         titleId = R.string.debug_open_mfkey32,
         onClick = { debugViewModel.openMfKey32(navController) }
     )
+    GrayDivider()
+    SwitchableElement(
+        titleId = R.string.experimental_application_catalog,
+        state = settings.applicationCatalog,
+        onSwitchState = debugViewModel::onSwitchApplicationCatalog
+    )
 }
