+++ components/settings/impl/src/main/java/com/flipperdevices/settings/impl/viewmodels/ExperimentalViewModel.kt
@@ -1,24 +1,19 @@
 package com.flipperdevices.settings.impl.viewmodels
 
-import androidx.datastore.core.DataStore
-import androidx.lifecycle.viewModelScope
 import androidx.navigation.NavController
 import com.flipperdevices.core.di.provideDelegate
-import com.flipperdevices.core.preference.pb.Settings
 import com.flipperdevices.core.ui.lifecycle.LifecycleViewModel
 import com.flipperdevices.filemanager.api.navigation.FileManagerEntry
 import com.flipperdevices.metric.api.MetricApi
 import com.flipperdevices.metric.api.events.SimpleEvent
 import com.flipperdevices.screenstreaming.api.ScreenStreamingFeatureEntry
-import kotlinx.coroutines.launch
 import tangle.viewmodel.VMInject
 import javax.inject.Provider
 
 class ExperimentalViewModel @VMInject constructor(
     fileManagerEntryProvider: Provider<FileManagerEntry>,
     metricApiProvider: Provider<MetricApi>,
-    private val screenStreamingFeatureEntry: ScreenStreamingFeatureEntry,
-    private val dataStore: DataStore<Settings>
+    private val screenStreamingFeatureEntry: ScreenStreamingFeatureEntry
 ) : LifecycleViewModel() {
     private val fileManagerEntry by fileManagerEntryProvider
     private val metricApi by metricApiProvider
@@ -31,15 +26,5 @@ class ExperimentalViewModel @VMInject constructor(
     fun onOpenScreenStreaming(navController: NavController) {
         metricApi.reportSimpleEvent(SimpleEvent.EXPERIMENTAL_OPEN_SCREEN_STREAMING)
         navController.navigate(screenStreamingFeatureEntry.ROUTE.name)
-    }
-
-    fun onSwitchApplicationCatalog(enabled: Boolean) {
-        viewModelScope.launch {
-            dataStore.updateData {
-                it.toBuilder()
-                    .setApplicationCatalog(enabled)
-                    .build()
-            }
-        }
     }
 }
